Checkpoint -- chart extraction is implemented and in testing
