+++ docker-compose.yaml
@@ -117,7 +117,8 @@ services:
     runtime: nvidia
 
   nv-ingest-ms-runtime:
-    image: nvcr.io/ohlfw0olaadg/ea-participants/nv-ingest:24.08
+    #image: nvcr.io/ohlfw0olaadg/ea-participants/nv-ingest:24.08
+    image: nv-ingest:devin_issue_fea_145
     build:
       context: ${NV_INGEST_ROOT:-.}
       dockerfile: "./Dockerfile"
@@ -141,7 +142,7 @@ services:
       - DEPLOT_INFER_PROTOCOL=http
       - DEPLOT_HTTP_ENDPOINT=https://ai.api.nvidia.com/v1/vlm/google/deplot
       - DOUGHNUT_GRPC_TRITON=triton-doughnut:8001
-      - INGEST_LOG_LEVEL=DEBUG
+      - INGEST_LOG_LEVEL=INFO
       - MESSAGE_CLIENT_HOST=redis
       - MESSAGE_CLIENT_PORT=6379
       - MINIO_BUCKET=${MINIO_BUCKET:-nv-ingest}
