+++ src/nv_ingest/schemas/ingest_job_schema.py
@@ -45,6 +45,7 @@ class TaskTypeEnum(str, Enum):
     store = "store"
     vdb_upload = "vdb_upload"
     table_data_extract = "table_data_extract"
+    chart_data_extract = "chart_data_extract"
 
 
 class FilterTypeEnum(str, Enum):
@@ -132,6 +133,8 @@ class IngestTaskVdbUploadSchema(BaseModelNoExt):
 class IngestTaskTableExtraction(BaseModelNoExt):
     params: Dict = {}
 
+class IngestChartTableExtraction(BaseModelNoExt):
+    params: Dict = {}
 
 class IngestTaskSchema(BaseModelNoExt):
     type: TaskTypeEnum
@@ -145,6 +148,7 @@ class IngestTaskSchema(BaseModelNoExt):
         IngestTaskFilterSchema,
         IngestTaskVdbUploadSchema,
         IngestTaskTableExtraction,
+        IngestChartTableExtraction
     ]
     raise_on_failure: bool = False
 
@@ -162,6 +166,7 @@ class IngestTaskSchema(BaseModelNoExt):
                 TaskTypeEnum.store: IngestTaskStoreSchema,
                 TaskTypeEnum.vdb_upload: IngestTaskVdbUploadSchema,
                 TaskTypeEnum.table_data_extract: IngestTaskTableExtraction,
+                TaskTypeEnum.chart_data_extract: IngestChartTableExtraction,
             }.get(task_type.lower())
 
             # logger.debug(f"Checking task_properties type for task type '{task_type}'")
