+++ src/nv_ingest/stages/nim/chart_extraction.py
@@ -2,34 +2,39 @@
 # All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 
-
 import logging
 import functools
-import traceback
+import pandas as pd
 from typing import Any
 from typing import Dict
+from typing import Optional
+from typing import Tuple
 
 from morpheus.config import Config
 
-from nv_ingest.schemas.metadata_schema import ContentTypeEnum
-from nv_ingest.schemas.table_extractor_schema import TableExtractorSchema
+from nv_ingest.schemas.chart_extractor_schema import ChartExtractorSchema
 from nv_ingest.stages.multiprocessing_stage import MultiProcessingBaseStage
-from nv_ingest.util.nim.helpers import call_image_inference_model
+from nv_ingest.util.image_processing.table_and_chart import join_cached_and_deplot_output
+from nv_ingest.util.nim.helpers import call_image_inference_model, create_inference_client
+from nv_ingest.util.image_processing.transforms import base64_to_numpy
 
 logger = logging.getLogger(f"morpheus.{__name__}")
 
 
-def _update_metadata(row, paddle_client, trace_info):
+def _update_metadata(row: pd.Series, cached_client: Any, deplot_client: Any, trace_info: Dict) -> Dict:
     """
-    Modifies the metadata of a row if the conditions for table extraction are met.
+    Modifies the metadata of a row if the conditions for chart extraction are met.
 
     Parameters
     ----------
     row : pd.Series
-        A row from the DataFrame containing metadata for the table extraction.
+        A row from the DataFrame containing metadata for the chart extraction.
 
-    paddle_client : Any
-        The client used to call the image inference model.
+    cached_client: Any
+        The client used to call the cached inference model.
+
+    deplot_client: Any
+        The client used to call the deplot inference model.
 
     trace_info : dict
         Trace information used for logging or debugging.
@@ -37,65 +42,105 @@ def _update_metadata(row, paddle_client, trace_info):
     Returns
     -------
     dict
-        The modified metadata.
+        The modified metadata if conditions are met, otherwise the original metadata.
+
+    Raises
+    ------
+    ValueError
+        If critical information (such as metadata) is missing from the row.
     """
-    metadata = row["metadata"]
 
-    base64_image = metadata["content"]
-    content_metadata = metadata["content_metadata"]
-    table_metadata = metadata.get("table_metadata")
+    metadata = row.get("metadata")
+    if metadata is None:
+        logger.error("Row does not contain 'metadata'.")
+        raise ValueError("Row does not contain 'metadata'.")
 
-    # Only modify if content type is structured and subtype is 'table' and table_metadata exists
-    if ((not content_metadata.type == ContentTypeEnum.STRUCTURED)
-            or (not content_metadata.subtype in ("chart",))
-            or (table_metadata is None)):
+    base64_image = metadata.get("content")
+    content_metadata = metadata.get("content_metadata", {})
+    chart_metadata = metadata.get("table_metadata")
+
+    # Only modify if content type is structured and subtype is 'chart' and chart_metadata exists
+    if ((content_metadata.get("type") != "structured")
+            or (content_metadata.get("subtype") != "chart")
+            or (chart_metadata is None)):
         return metadata
 
-    # Modify table metadata with the result from the inference model
-    table_metadata.table_content = call_image_inference_model(paddle_client, "paddle", base64_image, trace_info)
+    # Modify chart metadata with the result from the inference model
+    try:
+        image_array = base64_to_numpy(base64_image)
 
-    # Return the modified metadata to be updated in the DataFrame
+        deplot_result = call_image_inference_model(deplot_client, "deplot", image_array, trace_info=trace_info)
+        cached_result = call_image_inference_model(cached_client, "cached", image_array, trace_info=trace_info)
+        chart_content = join_cached_and_deplot_output(cached_result, deplot_result)
+
+        chart_metadata["table_content"] = chart_content
+    except Exception as e:
+        logger.error(f"Unhandled error calling image inference model: {e}", exc_info=True)
+        raise
+
     return metadata
 
 
-def _extract_chart_data(df, task_props, validated_config, trace_info=None):
+def _extract_chart_data(df: pd.DataFrame, task_props: Dict[str, Any],
+                        validated_config: Any, trace_info: Optional[Dict] = None) -> Tuple[pd.DataFrame, Dict]:
     """
-    Function to extract table data from a DataFrame.
+    Function to extract chart data from a DataFrame.
 
     Parameters
     ----------
     df : pd.DataFrame
-        DataFrame containing the content from which table data is to be extracted.
+        DataFrame containing the content from which chart data is to be extracted.
 
     task_props : dict
         Dictionary containing task properties and configurations.
 
-    validated_config : TableExtractorSchema
-        The validated configuration object for table extraction.
+    validated_config : Any
+        The validated configuration object for chart extraction.
 
-    trace_info : Optional[dict], default=None
-        Optional trace information for debugging or logging.
+    trace_info : dict, optional
+        Optional trace information for debugging or logging. Defaults to None.
+
+    Returns
+    -------
+    tuple
+        A tuple containing the updated DataFrame and the trace information.
+
+    Raises
+    ------
+    Exception
+        If any error occurs during the chart data extraction process.
     """
 
-    # TODO (Devin): Should be part of the stage_config
-    paddle_client = validated_config.paddle_endpoints
+    _ = task_props  # unused
+
+    deplot_client = create_inference_client(
+        validated_config.stage_config.deplot_endpoints,
+        validated_config.stage_config.auth_token,
+        validated_config.stage_config.deplot_infer_protocol
+    )
+
+    cached_client = create_inference_client(
+        validated_config.stage_config.cached_endpoints,
+        validated_config.stage_config.auth_token,
+        validated_config.stage_config.cached_infer_protocol
+    )
 
     if trace_info is None:
         trace_info = {}
+        logger.debug("No trace_info provided. Initialized empty trace_info dictionary.")
 
     try:
         # Apply the modify_metadata function to each row in the DataFrame
-        df["metadata"] = df.apply(_update_metadata, axis=1, args=(paddle_client, trace_info))
+        df["metadata"] = df.apply(_update_metadata, axis=1, args=(cached_client, deplot_client, trace_info))
 
         return df, trace_info
 
     except Exception as e:
-        traceback.print_exc()
-        logger.error(f"Error extracting table data: {e}")
+        logger.error("Error occurred while extracting chart data.", exc_info=True)
         raise
 
 
-def generate_table_extractor_stage(
+def generate_chart_extractor_stage(
         c: Config,
         stage_config: Dict[str, Any],
         task: str = "chart_data_extract",
@@ -103,7 +148,7 @@ def generate_table_extractor_stage(
         pe_count: int = 1,
 ):
     """
-    Helper function to generate a multiprocessing stage to perform table data extraction from PDF content.
+    Helper function to generate a multiprocessing stage to perform chart data extraction from PDF content.
 
     Parameters
     ----------
@@ -111,27 +156,27 @@ def generate_table_extractor_stage(
         Morpheus global configuration object.
 
     stage_config : dict
-        Configuration parameters for the table content extractor, passed as a dictionary
-        that will be validated against the `TableExtractorSchema`.
+        Configuration parameters for the chart content extractor, passed as a dictionary
+        that will be validated against the `ChartExtractorSchema`.
 
-    task : str, default="table_extraction"
-        The task name for the stage worker function, defining the specific table extraction process.
+    task : str, default="chart_extraction"
+        The task name for the stage worker function, defining the specific chart extraction process.
 
     task_desc : str, default="chart_data_extractor"
-        A descriptor used for latency tracing and logging during table extraction.
+        A descriptor used for latency tracing and logging during chart extraction.
 
     pe_count : int, default=1
-        The number of process engines to use for table data extraction. This value controls
+        The number of process engines to use for chart data extraction. This value controls
         how many worker processes will run concurrently.
 
     Returns
     -------
     MultiProcessingBaseStage
-        A configured Morpheus stage with an applied worker function that handles table data extraction
+        A configured Morpheus stage with an applied worker function that handles chart data extraction
         from PDF content.
     """
 
-    validated_config = TableExtractorSchema(**stage_config)
+    validated_config = ChartExtractorSchema(**stage_config)
     _wrapped_process_fn = functools.partial(_extract_chart_data, validated_config=validated_config)
 
     return MultiProcessingBaseStage(
